requirements.txt
# Audio synthesis
pyo>=1.0.5

# MIDI I/O
mido>=1.3.0

# OSC communication
python-osc>=1.8.3

# Scientific computing for feature extraction
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0

# Configuration and data
PyYAML>=6.0

# Development and testing
pytest>=7.4.0

